requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==3.2.15
django-colorfield==0.7.2
django-extra-fields==3.0.2
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djoser==2.1.0
drf-yasg==1.21.3
flake8==5.0.4
future==0.18.2
idna==3.3
importlib-metadata==1.7.0
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
mando==0.6.4
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.4.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.2.1
radon==5.1.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
serializers==0.2.4
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
zipp==3.8.1
